agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
cloudinary==1.41.0
Pillow==11.1.0
requests==2.32.0
transformers==4.47.0
torch==2.5.0
sentence-transformers==3.3.0
google-genai==1.73.1
mysql-connector-python==9.1.0
numpy==2.1.0
python-dotenv==1.0.0
